requirements.txt
absl-py==0.11.0
appdirs==1.4.4
arules==0.0.0
asgiref==3.2.3
astor==0.8.1
astroid==2.3.2
astunparse==1.6.3
audioread==2.1.9
azapi==3.0.4
beautifulsoup4==4.6.0
bs4==0.0.1
cachetools==4.0.0
certifi==2020.6.20
cffi==1.14.5
chardet==3.0.4
click==7.1.2
colorama==0.4.1
cycler==0.10.0
decorator==4.4.2
demjson==2.2.4
Django==3.0
django-cors-headers==3.7.0
django-crispy-forms==1.8.1
djangorestframework==3.12.2
findspark==1.4.2
Flask==1.1.2
Flask-Cors==3.0.10
Flask-Login==0.5.0
Flask-SQLAlchemy==2.4.4
flatbuffers==1.12
gast==0.3.3
google-auth==1.11.2
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
grpcio==1.32.0
gunicorn==20.0.4
h5py==2.10.0
idna==2.9
isort==4.3.21
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.17.0
Keras==2.4.3
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.2
kiwisolver==1.2.0
lazy-object-proxy==1.4.3
librosa==0.8.0
lightgbm==3.1.0
llvmlite==0.35.0
lxml==4.5.2
lyricsgenius==2.0.0
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.3.1
mccabe==0.6.1
mtcnn==0.1.0
nose==1.3.7
numba==0.52.0
numpy==1.19.5
oauthlib==3.1.0
opencv-python==4.3.0.36
opt-einsum==3.3.0
packaging==20.9
pandas==1.1.0
pickle-mixin==1.0.2
Pillow==7.2.0
pooch==1.3.0
protobuf==3.11.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pylint==2.4.3
pyparsing==2.4.7
python-dateutil==2.8.1
pytz==2020.1
PyYAML==5.3
requests==2.23.0
requests-oauthlib==1.3.0
resampy==0.2.2
rsa==4.0
scikit-learn==0.22.2.post1
scipy==1.4.1
seaborn==0.10.1
six==1.15.0
sklearn==0.0
SoundFile==0.10.3.post1
soupsieve==2.0.1
spotipy==2.16.0
SQLAlchemy==1.3.23
sqlparse==0.3.0
tensorboard==2.4.1
tensorboard-plugin-wit==1.8.0
tensorflow==2.1.0
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.1
termcolor==1.1.0
Theano==1.0.4+34.g473d74ea4
threadpoolctl==2.1.0
torch==1.7.1
typed-ast==1.4.0
typing-extensions==3.7.4.3
urllib3==1.25.8
Werkzeug==1.0.1
wordcloud==1.8.0
wrapt==1.12.1
xgboost==1.3.3
xlrd==1.2.0
xmltodict==0.12.0
